Simplify the `getSupplementaryPrivateUseRegExp` regular expression
I originally added an additional range to the regex in https://github.com/dequelabs/axe-core/pull/2102 while working on https://github.com/dequelabs/axe-core/issues/2101. The purpose of it was to match unicode characters in Supplementary Private Use Area A, and I got the regex itself from https://apps.timwhitlock.info/js/regex.

However, upon further review, it's clear that - for the most part - the added regex is the same as what was previously there.

The existing part was:

  [\uDB80-\uDBBF][\uDC00-\uDFFD]

The part I added (with an "or") was:

  [\uDB80-\uDBBE][\uDC00-\uDFFF]|\uDBBF[\uDC00-\uDFFD]

Note the overlap in the ranges. In fact, as far as I can tell, they're equivalent, except the previous part excluded anything where the low surrogate was DFFE or DFFF.

In this commit I've removed the unnecessary added portion after the "or", and fixed the original regex so that it correctly captures all characters in the range.